requirements.txt
streamlit==1.24.1
flask==3.1.1
flask-socketio==5.5.1
plotly==5.15.0
ccxt==1.106.100
python-telegram-bot==20.3
python-dotenv==1.0.0
ta==0.10.2
nest-asyncio==1.5.6
requests==2.31.0
pandas==2.0.3
numpy==1.26.4
scipy==1.11.1
matplotlib==3.8.0
websockets==11.0.3
